CHG: refactor, move all metadata functions into new file metadata.py
